requirements.txt
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5
numpy==1.20.1
Pillow>=8.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.0
six==1.15.0
tifffile==2020.9.3
icecream~=2.1.0